# Improve performance of from_unix_timestamp/to_unix_timestamp functions