… tests/queries/0_stateless/02812_from_to_utc_timestamp.reference